pyproject.toml
[tool.poetry]
name = "arxiv-browse"
version = "1.0.0"
description = "arXiv public browse and download app."
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"

Flask-SQLAlchemy = "~=2.4"  
MarkupSafe = "*"
SQLAlchemy = "~=1.4" 
bcrypt = "*"
cloudpathlib = {extras = ["gs"], version = "*"}
werkzeug = "~=2.3"
flask-rangerequest = "^0.0.2"
flask-s3 = "*"
flask-wtf = "*"
geoip2 = "*"

google-cloud-logging = "^3.2.5"
google-cloud-storage = "^2.5.0"
cloud-sql-python-connector = {extras = ["pymysql"], version = "^1.4.1"}

opentelemetry-api = "^1.13.0"
opentelemetry-exporter-gcp-trace = "^1.3.0"
opentelemetry-instrumentation-flask = "^0.34b0"
opentelemetry-instrumentation-requests = "^0.34b0"
opentelemetry-propagator-gcp = "^1.3.0"
opentelemetry-sdk = "^1.13.0"

mysqlclient = ">=2.1"
pydantic = "*"
python-dateutil = "*"
validators = "*"
types-python-dateutil = "^2.8.19.12"
pg8000 = "^1.30.1"
lxml = "^4.9.2"
xmltodict = "^0.13.0"
arxiv-base = {git = "https://github.com/arXiv/arxiv-base.git", rev = "1a6035ab"}
flask = "~=2.2.2"
google-cloud-compute = "^1.14.1"

[tool.poetry.dev-dependencies]
beautifulsoup4 = "*"
coverage = "*"
hypothesis = "*"
mypy = "*"
mypy-extensions = "*"
pycodestyle = "*"
pydocstyle = "*"
pylint = "*"
pytest = "*"
pytest-cov = "*"
pytest-easyread = "*"
pytest-html = "*"
pytest-mock = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
